perf: credit genesis sBTC balances directly instead of replaying a deposit per wallet (#2539)
* perf: credit genesis sBTC balances directly instead of replaying a deposit per wallet

* fix: run the genesis STX mint at the session epoch

* test: cover sub-dust sBTC funding and mint overflow

+++ components/clarity-repl/src/repl/session.rs
@@ -1588,7 +1588,7 @@ impl Session {
             return "Unable to parse the recipient address".red().to_string();
         };
 
-        let Ok(amount) = args[3].parse::<u64>() else {
+        let Ok(amount) = args[3].parse::<u128>() else {
             return "Unable to parse the amount".red().to_string();
         };
